evaluator: improve consistency

+++ evaluator/src/evaluator.rs
@@ -439,14 +439,14 @@ impl<'a> ExpressionEvaluator<'a> {
                 let v = self.eval_expr(expr, ts);
                 match (from, v) {
                     (UInt(_), Value::Unsigned(u)) => match to {
-                        UInt(_) => Value::Unsigned(u),
+                        UInt(_) => Value::Unsigned(u as u64),
                         Int(_) => Value::Signed(u as i64),
                         Float(_) => Value::new_float(u as f64),
                         _ => unimplemented!(),
                     },
                     (Int(_), Value::Signed(i)) => match to {
                         UInt(_) => Value::Unsigned(i as u64),
-                        Int(_) => Value::Signed(i),
+                        Int(_) => Value::Signed(i as i64),
                         Float(_) => Value::new_float(i as f64),
                         _ => unimplemented!(),
                     },
